marking cron for only production
