+++ packages/backend/convex/_generated/api.d.ts
@@ -13,8 +13,8 @@ import type * as analytics_analytics from "../analytics/analytics.js";
 import type * as audioProxy from "../audioProxy.js";
 import type * as auth from "../auth.js";
 import type * as batch_batch_coordination from "../batch/batch_coordination.js";
-import type * as batch_crons from "../batch/crons.js";
 import type * as batch_utils from "../batch/utils.js";
+import type * as crons from "../crons.js";
 import type * as everwhz from "../everwhz.js";
 import type * as everwhz_ai from "../everwhz_ai.js";
 import type * as http from "../http.js";
@@ -44,8 +44,8 @@ declare const fullApi: ApiFromModules<{
   audioProxy: typeof audioProxy;
   auth: typeof auth;
   "batch/batch_coordination": typeof batch_batch_coordination;
-  "batch/crons": typeof batch_crons;
   "batch/utils": typeof batch_utils;
+  crons: typeof crons;
   everwhz: typeof everwhz;
   everwhz_ai: typeof everwhz_ai;
   http: typeof http;
